Started implementation of delegate/event system that allows to pass parameters. Now the SystemWindow class has a mouse click event that the Application class subscribes to, and when triggered the Application's subscribed class recieves parameters of the mouse button that was clicked and the window ID where the click event occured And dynamically toggle wirframe ahsder and outline shader for objects in scene. The delegate/event system will be modified gradually.

+++ Source/Core/Graphics/Mesh.h
@@ -438,7 +438,7 @@ public:
 		Mesh ret_mesh;
 		float l_dist = 1.0f;
 		float l_span = 100.0f;
-		int		l_lineCount = ((l_span * 2) / l_dist) - 1;
+		int		l_lineCount = (int)(((l_span * 2) / l_dist) - 1);
 		int		l_vertexCount = l_lineCount * 4;
 
 		Vertex* l_pVertices = new Vertex[l_vertexCount];
